Authorize Middleware

+++ routes.php
@@ -1,20 +1,17 @@
 <?php
-
-use App\Controllers\UserController;
-
 $router->get('/', 'HomeController@index');
 $router->get('/listings', 'ListingController@index');
-$router->get('/listings/create', 'ListingController@create');
-$router->get('/listings/edit/{id}', 'ListingController@edit');
+$router->get('/listings/create', 'ListingController@create', ['auth']);
+$router->get('/listings/edit/{id}', 'ListingController@edit', ['auth']);
 $router->get('/listings/{id}', 'ListingController@show');
 
-$router->put('/listings/{id}', 'ListingController@update');
-$router->post('/listings', 'ListingController@store');
-$router->delete('/listings/{id}', 'ListingController@destroy');
+$router->put('/listings/{id}', 'ListingController@update', ['auth']);
+$router->post('/listings', 'ListingController@store', ['auth']);
+$router->delete('/listings/{id}', 'ListingController@destroy', ['auth']);
 
-$router->get('auth/register', 'UserController@create');
-$router->get('auth/login', 'UserController@login');
+$router->get('auth/register', 'UserController@create', ['guest']);
+$router->get('auth/login', 'UserController@login', ['guest']);
 
-$router->post('auth/register', "UserController@store");
-$router->post('auth/logout', "UserController@logout");
-$router->post('auth/login', "UserController@authenticate");
+$router->post('auth/register', "UserController@store", ['guest']);
+$router->post('auth/logout', "UserController@logout", ['auth']);
+$router->post('auth/login', "UserController@authenticate", ['guest']);
